satellite-processing-system/requirements.txt
# 六階段處理系統核心依賴
# 基於 netstack/requirements.txt 精簡版本

# 🚀 核心計算庫
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# 🛰️ 天體力學與軌道計算
skyfield>=1.45
sgp4>=2.20
pyephem>=4.1.0

# 📊 數據處理
PyYAML>=6.0
pydantic>=2.0.0

# 🌐 HTTP 與 API
requests>=2.31.0
httpx>=0.25.0

# 📁 文件處理
Pillow>=10.0.0

# 🕐 時間處理
python-dateutil>=2.8.0
pytz>=2023.3

# 🛠️ 系統工具
psutil>=5.9.0
tqdm>=4.65.0

# 📊 可選：數據庫支援 (如需要)
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# 🧪 開發與測試 - TDD/BDD 完整測試套件
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0                    # 測試覆蓋率
pytest-benchmark>=4.0.0              # 性能基準測試
pytest-bdd>=7.0.0                    # BDD 場景測試 (Phase 2)
pytest-html>=4.0.0                   # HTML 測試報告
pytest-json-report>=1.5.0            # JSON 測試報告
allure-pytest>=2.13.0                # 詳細測試報告 (Phase 2)

# 📈 日誌與監控
structlog>=23.1.0
rich>=13.5.0

# 🔒 配置管理
python-dotenv>=1.0.0

# 🌐 網路工具 (可選)
aiohttp>=3.8.0

# 📊 數值計算增強
matplotlib>=3.7.0
seaborn>=0.12.0

# 🚀 階段二性能優化套件 (GPU加速 + CPU並行)
# 注意：這些套件在無GPU環境下會自動降級到CPU實現，不會造成問題

# GPU計算庫 (可選，無GPU時自動使用CPU)
cupy-cuda12x>=12.0.0; platform_machine=="x86_64"     # CUDA 12.x (主要)
cupy-cuda11x>=11.0.0; platform_machine=="x86_64"     # CUDA 11.x (備選)

# JIT編譯器 (CPU/GPU通用)
numba>=0.58.0                       # 支援CPU和GPU加速

# 並行處理增強
multiprocessing-logging>=0.3.4      # 多進程日誌管理

# 性能分析工具
memory-profiler>=0.60.0             # 記憶體使用分析
line-profiler>=4.0.0                # 代碼性能分析
py-spy>=0.3.14                      # 低開銷性能分析器

# 測試增強 (新增)
pytest-xdist>=3.1.0                 # 並行測試執行